src/lessons/eric/requirements.txt
black==22.10.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
environs==9.5.0
idna==3.3
marshmallow==3.19.0
mypy-extensions==0.4.3
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pathspec==0.10.2
platformdirs==2.5.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
